Removes warnings: 	Size_t to int conversion. 	Use of incomplete type (forward of reference) 	Unused parameters

+++ include/fastrtps/xmlparser/XMLTree.h
@@ -97,7 +97,7 @@ class BaseNode
         parent_ = parent;
     }
 
-    int getNumChildren() const
+    size_t getNumChildren() const
     {
         return children.size();
     }
